refactor(loom): break the module dependency cycles
Loom's modules were already layered, but nothing enforced it: nine modules
reached back up into `web`, `acp` reached up into the review delivery worker
and the monitor loop, and the resulting cycles made the crate indivisible.

Four moves, no behaviour change:

- `lifecycle` takes the session state transitions — archive, adopt, warm
  session creation and their 16-function helper closure — out of the HTTP
  handlers that owned them. The callers that drive a session are mostly not
  requests (the reaper, the GitHub merge path, the restart adopt sweep, the
  watch engine), so they no longer reach up into `web`. Errors become plain
  `anyhow::Error`; the REST adapter maps them to a status. No caller lost
  information: every one already discarded the status or read `.message()`.

- `review_inbox` takes the claim/turn state machine out of `review_delivery`.
  It touches only the database and the ACP registry, which is exactly what
  `acp` was reaching up for; the GitHub side of delivery stays behind.

- `status` takes lifecycle promotion out of `monitor`. Translating a work-cycle
  signal into status, tags and an event is the same work whether the signal
  came from a `weaver hook` or an ACP turn boundary.

- `links` takes the dashboard deep-link formatters out of `web`. The Slack and
  GitHub integrations post these links without going near a request.

Verified by static analysis: the module graph now has zero upward edges against
a ctx / domain / ops / web layering.
