+++ crates/loom/src/lib.rs
@@ -32,6 +32,8 @@ pub(crate) mod history;
 pub mod ide;
 pub(crate) mod launch;
 pub mod launch_gate;
+pub mod lifecycle;
+pub mod links;
 pub mod logs;
 pub mod loom_config;
 pub mod mcp;
@@ -42,6 +44,7 @@ pub(crate) mod provision;
 pub mod repo;
 pub mod repo_env;
 pub mod review_delivery;
+pub mod review_inbox;
 pub mod runner;
 pub mod runs;
 pub(crate) mod runtime;
@@ -53,6 +56,7 @@ pub mod session_manager;
 pub(crate) mod setup;
 pub mod shell;
 pub(crate) mod slack;
+pub mod status;
 pub(crate) mod tasks;
 pub(crate) mod terminal;
 pub mod user_token;
